requirements.txt
Flask==2.3.2
gunicorn==20.1.0
langchain==0.0.247
openai==0.27.8
unstructured==0.8.7
faiss-cpu==1.7.4
python-dotenv==1.0.0
tiktoken==0.4.0
pgvector
pandas
numpy
psycopg2
psycopg2cffi

